Add files via upload
This project is based on Backend REST API with CRUD operations

To Run this project Clone it and install modules using

npm install

Then Create .env file and create MONGO Variable and specify Value. That's it. You are ready to go. To execute this project just type

npm start.

Thank you!